# openapi.yaml
openapi: 3.1.0

info:
  title: hiyocord Nexus
  version: 1.0.0
  description: API documentation for hiyocord Nexus, a service for managing Discord interactions.

servers:
  - url: https://nexus.hiyocord.org
    description: Official Production server

paths:
  /interactions:
    description: "Discord interaction"

  /api/manifests:
    description: "Get list of registered manifests"
    get:
      tags:
        - Web API
      summary: "List all registered manifests"
      responses:
        '200':
          description: "Successfully retrieved manifests list"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ManifestLatestVersion'
        default:
          description: "Server error"

    post:
      description: "Endpoint to submit manifest from service workers"
      requestBody:
        required: true
        content:
          application/json:
            schema:
                $ref: '#/components/schemas/ManifestAnyVersion'
      responses:
        '200':
          description: "ok"
        '400':
          description: "Invalid input data"
        default:
          description: "Server error"

  /api/manifests/{id}:
    description: "Get or delete a specific manifest"
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: "Manifest ID"
    get:
      tags:
        - Web API
      summary: "Get manifest details"
      responses:
        '200':
          description: "Successfully retrieved manifest details"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManifestLatestVersion'
        '404':
          description: "Manifest not found"
        default:
          description: "Server error"
    delete:
      tags:
        - Web API
      summary: "Delete a manifest"
      responses:
        '200':
          description: "Successfully deleted manifest"
        '404':
          description: "Manifest not found"
        default:
          description: "Server error"

  /.well-known/nexus-public-key:
    description: "Endpoint to retrieve Nexus public key for signature verification"
    get:
      responses:
        '200':
          description: "Successfully retrieved Nexus public key"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NexusPublicKey'
        default:
          description: "Server error"

  /api/proxy/discord/v10:
    summary: discord api v10 proxy
    description: schema ref discord api v10

  /api/auth/discord/authorize:
    description: "Discord OAuth2 authorization redirect endpoint"
    get:
      tags:
        - Web API
      summary: "Redirect to Discord OAuth2 authorization page"
      parameters:
        - name: origin
          in: query
          required: true
          schema:
            type: string
          description: "Frontend origin (window.location.origin)"
      responses:
        '302':
          description: "Redirect to Discord OAuth2 authorization page"
        '400':
          description: "Invalid origin parameter"
        default:
          description: "Server error"

  /api/auth/discord:
    description: "Discord OAuth2 authentication endpoint"
    post:
      tags:
        - Web API
      summary: "Exchange Discord OAuth2 code for access token"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                  description: "OAuth2 authorization code from Discord"
                redirect_uri:
                  type: string
                  description: "Redirect URI used in the OAuth2 flow"
              required:
                - code
                - redirect_uri
      responses:
        '200':
          description: "Successfully authenticated. JWT is set in HttpOnly Cookie."
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    description: "Discord user information"
                    properties:
                      id:
                        type: string
                      username:
                        type: string
                      discriminator:
                        type: string
                      avatar:
                        type: string
        '401':
          description: "Authentication failed"
        default:
          description: "Server error"

  /api/auth/me:
    description: "Get current authenticated user"
    get:
      tags:
        - Web API
      summary: "Get current user information"
      responses:
        '200':
          description: "Successfully retrieved user information"
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        description: "Discord User ID"
        '401':
          description: "Unauthorized"
        default:
          description: "Server error"

  /api/auth/logout:
    description: "Logout and clear authentication cookie"
    post:
      tags:
        - Web API
      summary: "Logout user"
      responses:
        '200':
          description: "Successfully logged out. Cookie is cleared."
          content:
            application/json:
              schema:
                type: object
        default:
          description: "Server error"

components:
  schemas:
    NexusPublicKey:
      type: object
      properties:
        algorithm:
          type: string
          description: "Public key signature algorithm"
          enum:
            - ed25519
            - ecdsa-p256
            - rsa-pss-2048
        public_key:
          type: string
          description: "Base64-encoded Nexus public key"
      required:
        - algorithm
        - public_key

    ManifestLatestVersion:
      type: object
      $ref: '#/components/schemas/Manifest_V1'

    ManifestAnyVersion:
      type: object
      oneOf:
        - $ref: '#/components/schemas/Manifest_V1'
      required:
        - version

    Manifest_V1:
      type: object
      properties:
        version:
          type: string
          description: "Manifest schema version."
          pattern: "^1\\.[0-9]\\.[0-9]$"
        id:
          type: string
          description: "Unique identifier for the application"
          example: "nexus.hiyocord.org"
        name:
          type: string
          description: "Name of the application"
        icon_url:
          type:
            - string
            - "null"
          format: uri
          description: "URL to the application's icon"
        description:
          type: string
          description: "Description of the application"
        base_url:
          type: string
          format: uri
          description: "Base URL for the application"
        application_commands:
          type: object
          description: "Application commands supported by the application"
          required:
            - global
            - guild
          properties:
            global:
              type: array
              description: "Global application commands"
              items:
                $ref: '#/components/schemas/DiscordCommand'
            guild:
              type: array
              description: "Guild-specific application commands"
              items:
                type: object
                allOf:
                  - $ref: '#/components/schemas/DiscordCommand'
                  - type: object
                    properties:
                      guild_id:
                        type: array
                        description: "List of guild IDs where the command is available"
                        items:
                          type: string
                    required:
                      - guild_id
        message_component_ids:
          type: array
          items:
            type: string
        modal_submit_ids:
          type: array
          items:
            type: string
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
        signature_algorithm:
          type: string
          description: "Public key signature algorithm for authentication"
          enum:
            - ed25519
            - ecdsa-p256
            - rsa-pss-2048
          default: ed25519
        public_key:
          type: string
          description: "Base64-encoded public key for verifying requests from this service worker"
      required:
        - version
        - id
        - name
        - base_url
        - application_commands
        - message_component_ids
        - modal_submit_ids
        - permissions
        - signature_algorithm
        - public_key

    Permission:
      oneOf:
        - type: object
          required:
            - type
          properties:
            type:
              type: string
              enum:
                - DISCORD_BOT
        - type: object
          properties:
            type:
              type: string
              const: DISCORD_API_SCOPE
            scopes:
              type: object
              description: "許可するDiscord APIエンドポイントとHTTPメソッドの組み合わせ"
              additionalProperties:
                type: array
                items:
                  type: string
              example:
                "/channels/:channel_id/messages": ["POST"]
                "/channels/:channel_id/messages/:message_id": ["GET", "PATCH", "DELETE"]
                "/guilds/:guild_id/members/:user_id": ["GET"]
          required:
            - type
            - scopes

    DiscordCommand:
      type: object
      allOf:
        - $ref: "https://raw.githubusercontent.com/discord/discord-api-spec/refs/heads/main/specs/openapi.json#/components/schemas/ApplicationCommandCreateRequest"
      description: "A Discord application command object"

    DiscordInteraction:
      description: "A Discord interaction object"
      type: object
      properties:
        type:
          type: integer
          description: "Type of interaction"
        version:
          type: integer
          description: "Version of the interaction"
        application_id:
          type: string
          description: "ID of the application"
        event:
          type: object
          description: "Event data"
          properties:
            data:
              type: object
              description: "Interaction data"

    ManifestSummary:
      type: object
      description: "Summary information of a registered manifest"
      properties:
        id:
          type: string
          description: "Manifest ID"
        name:
          type: string
          description: "Manifest name"
        icon_url:
          type:
            - string
            - "null"
          format: uri
          description: "URL to the application's icon"
        description:
          type: string
          description: "Manifest description"
        base_url:
          type: string
          format: uri
          description: "Base URL of the service worker"
        command_count:
          type: object
          description: "Number of registered commands"
          properties:
            global:
              type: integer
              description: "Number of global commands"
            guild:
              type: integer
              description: "Number of guild-specific commands"
        component_count:
          type: integer
          description: "Number of message component IDs"
        modal_count:
          type: integer
          description: "Number of modal submit IDs"
      required:
        - id
        - name
        - description
        - base_url
        - command_count
        - component_count
        - modal_count